Whole Brain Circuitry Profiling/CircuitFound/requirements.txt
absl-py==1.3.0
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
astunparse==1.6.3
attrs==23.1.0
axial-attention==0.6.1
backcall==0.2.0
batchgenerators==0.24
beautifulsoup4==4.12.2
bleach==6.0.0
brotlipy==0.7.0
cachetools==5.2.0
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi @ file:///tmp/abs_98z5h56wf8/croots/recipe/cffi_1659598650955/work
charset-normalizer==2.1.1
click==8.1.7
comm==0.1.3
connected-components-3d==3.10.3
contourpy==1.0.6
cryptography==3.3.2
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dicom2nifti==2.4.5
docker-pycreds==0.4.0
einops==0.6.0
executing==1.2.0
fastjsonschema==2.16.3
filelock==3.12.4
flatbuffers==22.10.26
fonttools==4.38.0
fqdn==1.5.1
fsspec==2023.9.1
future==0.18.2
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.36
google-auth==2.14.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.50.0
h5py==3.7.0
huggingface-hub==0.17.2
idna @ file:///croot/idna_1666125576474/work
imageio==2.22.4
imgaug==0.4.0
importlib-metadata==5.0.0
importlib-resources==5.12.0
ipykernel==6.23.1
ipython==8.12.2
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
jedi==0.18.2
jeepney==0.8.0
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
keras==2.11.0
keyring==23.4.1
keyrings.alt==3.1
kiwisolver==1.4.4
libclang==14.0.6
libtiff==0.4.2
linecache2==1.0.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
MedPy==0.4.0
mistune==2.0.5
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
mkl-service==2.4.0
monai==1.2.0
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.4.0
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==2.8.8
nibabel==4.0.2
# Editable Git install with no remote (nnunet==1.7.0)
-e /media/root/e8449930-91ce-40a4-a5c5-87d4d2cd1568/lpq/nnUNet/D-LMBMap_nnUNet
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.23.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.1
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pkgutil_resolve_name==1.3.10
platformdirs==3.5.1
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==3.19.6
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydicom==2.3.1
Pygments==2.15.1
pylibtiff==0.5.1
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil==2.8.2
python-gdcm==3.0.19
python-json-logger==2.0.7
pytz==2022.6
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.0.2
qtconsole==5.4.3
QtPy==2.3.1
requests @ file:///opt/conda/conda-bld/requests_1657734628632/work
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
safetensors==0.3.3
scikit-image==0.19.3
scikit-learn==1.1.3
scipy==1.9.3
SecretStorage==3.3.3
Send2Trash==1.8.2
sentry-sdk==1.31.0
setproctitle==1.3.2
Shapely==1.8.5.post1
SimpleITK==2.2.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
sklearn==0.0.post1
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.4.1
stack-data==0.6.2
synapseclient==3.0.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.1.1
terminado==0.17.1
threadpoolctl==3.1.0
tifffile==2022.10.10
timm==0.9.7
tinycss2==1.2.1
torch==1.11.0
torchaudio==0.11.0
torchprofile==0.0.4
torchsummary==1.5.1
torchvision==0.12.0
tornado==6.3.2
tqdm==4.64.1
traceback2==1.4.0
traitlets==5.9.0
typing_extensions==4.4.0
unittest2==1.1.0
uri-template==1.2.0
urllib3 @ file:///croot/urllib3_1666298941550/work
wandb==0.15.10
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.2
widgetsnbextension==4.0.7
wrapt==1.14.1
zipp==3.10.0
